Optionally allow the client to trigger key generation
Client-initiated key generation normally should not be needed, so make
it optional. When allowed, the client can specify key parameters (key
type, length, UID etc) and also decide whether it should be protected
with a passphrase. But if chosen to use a passphrase, it needs to be
provided with pinentry running in the backend - client passphrase is not
allowed.

This option is disabled by default and there is no provision to enable
it (yet).