server: Use stdin/-out instead of bi-directional socket
When we are spawned directly from qrexec-agent instead of through
qrexec-fork-server we are running as a different user than qrexec-agent
and therefore can't sent SIGUSR1 to request switching to a bidirectional
socket.

To avoid to have two code paths always use stdin/-out for now. We still
can revert this if there's a better solution.