Prevent gpg from starting local agent
Wrap gpg-agent on the client side in a script that is no-op if
'split-gpg2-clieng' service is enabled. This avoids gpg overriding
split-gpg2's agent socket if user is slow to accept a qrexec call.

Use `gpgconf --list-components` and then exec in the wrapper instead of
`gpgconf --launch`, to preserve program arguments (`gpgconf --launch`
allows to set only some of them, skipping others may result in
weird behavior).