Allow an empty passphrase to be specified
The simplest way to non-interactively create a key with no passphrase is
as follows:

$ gpg2 --quick-generate-key --batch --passphrase ''

Unfortunately, this causes GnuPG to use --inq-passwd, which is blocked
by split-gpg2.  As a result, this command fails with a very confusing
"key generation failed: End of file" error.

To solve this problem, allow the client to use --inq-passwd, but when
gpg-agent responds with a NEWPASSWD inquire, only allow END as the
reply.  This makes commands like the above work, while still forbidding
specifying a non-empty passphrase.